Hint button functionality + more
Hint button now randomly selects a cell and fills it in, adding 1 minute to the timer. If the puzzle is already solved, win is pushed through. addTime() calls implemented in checkBoard and  giveHint, and a new function has been added to replace the code in checkBoard so that checking that the board is complete can be called by other functions/actions.